requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
Flask==2.3.2
frozenlist==1.5.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.31.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
yarl==1.18.3
